# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SiteDrop - HTML site hosting platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Site {
  id: number;
  name: string;
  title: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  htmlContent?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface SiteInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  name: string;
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minLength 1 */
  htmlContent: string;
}

export interface NameCheckResult {
  available: boolean;
  name: string;
}

export interface SiteStats {
  totalSites: number;
  recentSites: Site[];
}

export interface GenerateSiteInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  name: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 10 */
  description: string;
}

export interface ErrorResponse {
  error: string;
}