# lib/api-zod/src/generated/types/generateSiteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SiteDrop - HTML site hosting platform API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateSiteInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  name: string;
  /** @minLength 1 */
  title: string;
  /** @minLength 10 */
  description: string;
}